requirements.txt
Flask>=2.3
python-dotenv>=1.0
twilio>=8.0
requests>=2.31
google-api-python-client>=2.100
google-auth>=2.23
google-auth-httplib2>=0.1.0
